pyproject.toml
[project]
name = "zomma-kg"
version = "0.3.0"
description = "Embedded knowledge graph library for document understanding"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "Zomma Labs"}]
keywords = [
    "knowledge-graph",
    "rag",
    "llm",
    "financial-documents",
    "document-understanding",
    "embedded-database",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database :: Database Engines/Servers",
    "Typing :: Typed",
]

dependencies = [
    # Storage
    "duckdb>=0.10",
    "lancedb>=0.6",
    "pyarrow>=14.0",
    "filelock>=3.12",
    # Core utilities
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Async
    "anyio>=4.0",
    # Types
    "typing-extensions>=4.8",
    "python-dotenv>=1.2.1",
    # Scientific computing (for entity deduplication)
    "numpy>=1.20",
    "scipy>=1.10",
    # CLI
    "typer[all]>=0.9.0",
]

[project.optional-dependencies]
# LLM providers (LangChain-based)
openai = [
    "langchain-openai>=0.1",
    "langchain-core>=0.2",
    "tiktoken>=0.5",
]
anthropic = [
    "anthropic>=0.18",
]
google = [
    "google-genai>=1.0",
]

# Embedding providers
voyage = [
    "voyageai>=0.2",
]
local-embed = [
    "sentence-transformers>=2.2",
    "torch>=2.0",
]

# MCP Server
mcp = [
    "mcp[cli]>=1.2.0",
]

# Bundles
cloud = [
    "zomma-kg[openai,anthropic,google,voyage]",
]
all = [
    "zomma-kg[cloud,mcp]",
]
full = [
    "zomma-kg[all,local-embed]",
]

# Development
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
zomma-kg = "zomma_kg.cli:main"

[project.urls]
Homepage = "https://github.com/Zomma-Labs/zomma-kg"
Documentation = "https://zomma-kg.readthedocs.io"
Repository = "https://github.com/Zomma-Labs/zomma-kg"
Issues = "https://github.com/Zomma-Labs/zomma-kg/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["zomma_kg"]

[tool.hatch.build.targets.sdist]
include = [
    "zomma_kg/**/*.py",
    "zomma_kg/py.typed",
    "zomma_kg/data/**/*.json",
    "LICENSE",
    "README.md",
]

[tool.hatch.build.targets.wheel.force-include]
"zomma_kg/data" = "zomma_kg/data"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
